+++ packages/ai/core/generate-object/output-strategy.ts
@@ -10,12 +10,13 @@ import {
 import { safeValidateTypes, ValidationResult } from '@ai-sdk/provider-utils';
 import { asSchema, DeepPartial, Schema } from '@ai-sdk/ui-utils';
 import { z } from 'zod';
+import { NoObjectGeneratedError } from '../../errors/no-object-generated-error';
 import {
   AsyncIterableStream,
   createAsyncIterableStream,
 } from '../util/async-iterable-stream';
-import { NoObjectGeneratedError } from './no-object-generated-error';
 import { ObjectStreamPart } from './stream-object-result';
+import { LanguageModelResponseMetadata, LanguageModelUsage } from '../types';
 
 export interface OutputStrategy<PARTIAL, RESULT, ELEMENT_STREAM> {
   readonly type: 'object' | 'array' | 'enum' | 'no-schema';
@@ -35,7 +36,14 @@ export interface OutputStrategy<PARTIAL, RESULT, ELEMENT_STREAM> {
     partial: PARTIAL;
     textDelta: string;
   }>;
-  validateFinalResult(value: JSONValue | undefined): ValidationResult<RESULT>;
+  validateFinalResult(
+    value: JSONValue | undefined,
+    context: {
+      text: string;
+      response: LanguageModelResponseMetadata;
+      usage: LanguageModelUsage;
+    },
+  ): ValidationResult<RESULT>;
 
   createElementStream(
     originalStream: ReadableStream<ObjectStreamPart<PARTIAL>>,
@@ -52,9 +60,22 @@ const noSchemaOutputStrategy: OutputStrategy<JSONValue, JSONValue, never> = {
 
   validateFinalResult(
     value: JSONValue | undefined,
+    context: {
+      text: string;
+      response: LanguageModelResponseMetadata;
+      usage: LanguageModelUsage;
+    },
   ): ValidationResult<JSONValue> {
     return value === undefined
-      ? { success: false, error: new NoObjectGeneratedError() }
+      ? {
+          success: false,
+          error: new NoObjectGeneratedError({
+            message: 'No object generated: response did not match schema.',
+            text: context.text,
+            response: context.response,
+            usage: context.usage,
+          }),
+        }
       : { success: true, value };
   },