+++ packages/google/src/google-generative-ai-language-model.ts
@@ -120,6 +120,7 @@ export class GoogleGenerativeAILanguageModel implements LanguageModelV1 {
         const { tools, toolConfig, toolWarnings } = prepareTools(
           mode,
           this.settings.useSearchGrounding ?? false,
+          this.modelId.includes('gemini-2'),
         );
 
         return {
@@ -246,6 +247,7 @@ export class GoogleGenerativeAILanguageModel implements LanguageModelV1 {
       providerMetadata: {
         google: {
           groundingMetadata: candidate.groundingMetadata ?? null,
+          safetyRatings: candidate.safetyRatings ?? null,
         },
       },
       request: { body },
@@ -327,6 +329,7 @@ export class GoogleGenerativeAILanguageModel implements LanguageModelV1 {
               providerMetadata = {
                 google: {
                   groundingMetadata: candidate.groundingMetadata ?? null,
+                  safetyRatings: candidate.safetyRatings ?? null,
                 },
               };
             }
@@ -443,13 +446,33 @@ const contentSchema = z.object({
   ),
 });
 
+// https://ai.google.dev/gemini-api/docs/grounding
 // https://cloud.google.com/vertex-ai/generative-ai/docs/multimodal/ground-gemini#ground-to-search
 export const groundingMetadataSchema = z.object({
   webSearchQueries: z.array(z.string()).nullish(),
+  retrievalQueries: z.array(z.string()).nullish(),
   searchEntryPoint: z
     .object({
       renderedContent: z.string(),
     })
+    .nullish(),
+  groundingChunks: z
+    .array(
+      z.object({
+        web: z
+          .object({
+            uri: z.string(),
+            title: z.string(),
+          })
+          .nullish(),
+        retrievedContext: z
+          .object({
+            uri: z.string(),
+            title: z.string(),
+          })
+          .nullish(),
+      }),
+    )
     .nullish(),
   groundingSupports: z
     .array(
@@ -459,8 +482,11 @@ export const groundingMetadataSchema = z.object({
           endIndex: z.number().nullish(),
           text: z.string().nullish(),
         }),
-        groundingChunkIndices: z.array(z.number()),
-        confidenceScores: z.array(z.number()),
+        segment_text: z.string().nullish(),
+        groundingChunkIndices: z.array(z.number()).nullish(),
+        supportChunkIndices: z.array(z.number()).nullish(),
+        confidenceScores: z.array(z.number()).nullish(),
+        confidenceScore: z.array(z.number()).nullish(),
       }),
     )
     .nullish(),
@@ -474,11 +500,22 @@ export const groundingMetadataSchema = z.object({
     .nullish(),
 });
 
+// https://cloud.google.com/vertex-ai/generative-ai/docs/multimodal/configure-safety-filters
+export const safetyRatingSchema = z.object({
+  category: z.string(),
+  probability: z.string(),
+  probabilityScore: z.number().nullish(),
+  severity: z.string().nullish(),
+  severityScore: z.number().nullish(),
+  blocked: z.boolean().nullish(),
+});
+
 const responseSchema = z.object({
   candidates: z.array(
     z.object({
       content: contentSchema.nullish(),
       finishReason: z.string().nullish(),
+      safetyRatings: z.array(safetyRatingSchema).nullish(),
       groundingMetadata: groundingMetadataSchema.nullish(),
     }),
   ),
@@ -499,6 +536,7 @@ const chunkSchema = z.object({
       z.object({
         content: contentSchema.nullish(),
         finishReason: z.string().nullish(),
+        safetyRatings: z.array(safetyRatingSchema).nullish(),
         groundingMetadata: groundingMetadataSchema.nullish(),
       }),
     )