+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,36 @@
 
 ### Patch Changes
 
+- Updated dependencies [304e6d3]
+- Updated dependencies [304e6d3]
+  - ai@4.0.18
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [54bbf21]
+  - ai@4.0.17
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [e3fac3f]
+  - ai@4.0.16
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [cc16a83]
+- Updated dependencies [cc16a83]
+  - ai@4.0.15
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [09a9cab]
   - ai@4.0.14